Remove renderApp file in favour of renderRoute from test-utils
