+++ app/javascript/src/views/BookingSetup/__tests__/index.js
@@ -1,4 +1,4 @@
-import renderApp from "../../../testHelpers/renderApp";
+import { renderRoute } from "test-utils";
 import generateTypes from "../../../__mocks__/graphqlFields";
 import VIEWER from "../../../graphql/queries/viewer";
 import GET_SETUP_DATA from "../getSetupData";
@@ -25,7 +25,7 @@ test("Renders the project type step if payments are setup", async () => {
     specialist,
   });
 
-  const app = renderApp({
+  const app = renderRoute({
     route: "/book/rec1234",
     graphQLMocks: [
       {
